Initial struct support, remove extraneous Type constructor.
The Type constructor was removed, because otherwise struct declarations
would be wrapped inside a Type constructor as well, which would look
somewhat weird.

As the InitDeclarator is optional, it should be 0 or more, rather than 1
or more times (so * instead of +). This however caused an ambiguity in
the case of a declaration like "a b;", where we have to check whether b
is a defined type or not, otherwise, it could be, according to the
grammar refer to a type, instead of the variable name.