Fix #include module imports.
Now the `Program(filename, _)` filename will be the absolute path,
prefixed with some eclipse:///project stuff.

It gets the context location URL, which is the VFS URL of the project.
This first part is than stripped from the file VFS URL. Finally using
the <project-path> strategy, the absolute path of the file is generated.

It is kinda hacky, but hopefully more robust than just stripping of the
first characters to get rid of the 'eclipse://'.