Improve Writing/Reading channels for ints, structs and pointers of structs
Create a temporary variable of the type of the channel (only if necessary,
when the channel is not an int-type or the expression is not a variable
reference).

Now the semantics are similar to calling functions. When you call a
function like:

  struct x a;
  f(a);

in C, the struct is copied too, unless you pass a pointer:

  struct x a;
  f(&a);

The same holds for writing to channels, as you can have a channel of
`chan<struct x *>`, where you pass pointers of structs to the other
side. If you want to write to a channel by value, you create a channel
of `chan<struct x>`.